requirements.txt
# Core dependencies
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Configuration & Logging
pyyaml>=6.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# Deep Learning (for later stages)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Utilities
scikit-learn>=1.3.0
scikit-image>=0.21.0
pandas>=2.0.0
